% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_occupancies.R
\name{calc_occupancies}
\alias{calc_occupancies}
\title{Calculate multistate occupancy times}
\usage{
calc_occupancies(
  p_list,
  age = NULL,
  init = NULL,
  origin_state = NULL,
  delim = NULL,
  age_interval = 1,
  age_col = "age",
  trans_col = "from_to",
  p_col = "p"
)
}
\arguments{
\item{p_list}{A named list of age-specific transition probabilities. Recall data.frames and tibbles can function as lists too.}

\item{age}{A numeric vector of age values.}

\item{init}{Optional named vector of initial state composition.}

\item{origin_state}{Optional single origin state (alternative to \code{init}).}

\item{delim}{Character delimiter separating from/to in transition names.}

\item{age_interval}{integer value representing time step (default 1)}

\item{age_col}{Name of the age column (default = \code{"age"}) when input is a data.frame.}

\item{trans_col}{Name of the transition column (default = \code{"from_to"}) when input is a tidy \code{data.frame}.}

\item{p_col}{Name of the transition probability column (default = \code{"p"}) when input is a tidy \code{data.frame}.}
}
\value{
A data frame of occupancy times by state and age.
}
\description{
Calculates state occupancy over age using a list of transition probabilities.
}
\examples{

prep <- prepare_p_list(transitions_example)
occup <- calc_occupancies(prep$p_list, prep$age, origin_state = "P")
head(occup)
# or without prep (but with prep overhead happening internally)
occup <- calc_occupancies(transitions_example, origin_state = "P")
# it's ca 2x faster to have the data in list format in advance, FYI.
}
