% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_occupancies.R
\name{calc_occupancies}
\alias{calc_occupancies}
\title{Calculate multistate occupancy times}
\usage{
calc_occupancies(
  p_list,
  age,
  init = NULL,
  origin_state = NULL,
  delim = "->",
  age_interval = 1
)
}
\arguments{
\item{p_list}{A named list of age-specific transition probabilities. Recall data.frames and tibbles can function as lists too.}

\item{age}{A numeric vector of age values.}

\item{init}{Optional named vector of initial state composition.}

\item{origin_state}{Optional single origin state (alternative to \code{init}).}

\item{delim}{Character delimiter separating from/to in transition names.}

\item{age_interval}{integer value representing time step (default 1)}
}
\value{
A data frame of occupancy times by state and age.
}
\description{
Calculates state occupancy over age using a list of transition probabilities.
}
\examples{
Create p_list the tidy way
p_df <- transitions_example |>
 tidyr::pivot_wider(names_from = from_to, values_from = p) |>
 dplyr::arrange(age)
age <- p_df$age
p_list <- as.list(p_df[ , -1])
trans_info <- split_transitions(p_list)
from_states <- trans_info$from
to_states <- trans_info$to
p_list <- fill_missing_transitions(p_list, from_states, to_states) |>
infer_self_transitions(p_list, from_states, to_states)
out <- calc_occupancies(p_list, age, origin_state = "P")
head(out)
}
